Poprawa kodu
(cherry picked from commit 4c852c6612fffdb6077d5c8ff766ecb264cee266)

+++ src/main/java/pl/com/itsystems/cookbook/category/CategoryService.java
@@ -1,15 +1,20 @@
 package pl.com.itsystems.cookbook.category;
 
 import org.springframework.stereotype.Service;
+import pl.com.itsystems.cookbook.recipe.Recipe;
+import pl.com.itsystems.cookbook.recipe.RecipeRepository;
 
 import java.util.List;
 
 @Service
 public class CategoryService {
     private final CategoryRepository categoryRepository;
+    private final RecipeRepository recipeRepository;
+    private static final Long NO_CATEGORY_ID = 0L;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(CategoryRepository categoryRepository, RecipeRepository recipeRepository) {
         this.categoryRepository = categoryRepository;
+        this.recipeRepository = recipeRepository;
     }
 
     public List<Category> getAllCategories() {
@@ -25,6 +30,15 @@ public class CategoryService {
     }
 
     public void deleteCategory(Long id) {
-        categoryRepository.deleteById(id);
+        List<Recipe> recipes = recipeRepository.findAllByCategoryId(id);
+        Category noCategory = categoryRepository.findById(NO_CATEGORY_ID).orElse(null);
+        for (Recipe recipe : recipes) {
+            recipe.setCategory(noCategory);
+        }
+
+        recipeRepository.saveAll(recipes);
+        if (id > NO_CATEGORY_ID) {
+            categoryRepository.deleteById(id);
+        }
     }
 }
